Add new localizable strings for accessibility content description and adjust tests as a result.  Update picker subtext color.
