+++ libs/test/SalesforceSDKTest/src/com/salesforce/androidsdk/ui/PickerBottomSheetActivityTest.kt
@@ -29,17 +29,21 @@ package com.salesforce.androidsdk.ui
 import androidx.activity.ComponentActivity
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.ui.test.assertIsDisplayed
+import androidx.compose.ui.test.assertIsNotSelected
+import androidx.compose.ui.test.assertIsSelected
 import androidx.compose.ui.test.hasContentDescription
+import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.isDisplayed
 import androidx.compose.ui.test.isNotDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
+import androidx.compose.ui.test.onChild
 import androidx.compose.ui.test.onChildAt
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import com.salesforce.androidsdk.R
-import com.salesforce.androidsdk.ui.components.BACK_BUTTON_CD
-import com.salesforce.androidsdk.ui.components.CLOSE_BUTTON_CD
+import com.salesforce.androidsdk.config.LoginServerManager.LoginServer
 import com.salesforce.androidsdk.ui.components.PickerStyle
+import org.junit.Assert
 import org.junit.Rule
 import org.junit.Test
 
@@ -52,8 +56,23 @@ class PickerBottomSheetActivityTest {
     @get:Rule
     val androidComposeTestRule = createAndroidComposeRule<ComponentActivity>()
 
-    private val closeButtonCd = hasContentDescription(CLOSE_BUTTON_CD)
-    private val backButtonCd = hasContentDescription(BACK_BUTTON_CD)
+    private val serverPickerCd = hasContentDescription(
+        androidComposeTestRule.activity.getString(R.string.sf__server_picker_content_description)
+    )
+    private val accountPickerCd = hasContentDescription(
+        androidComposeTestRule.activity.getString(R.string.sf__account_picker_content_description)
+    )
+    private val closeButtonCd = hasContentDescription(
+        androidComposeTestRule.activity.getString(R.string.sf__server_close_button_content_description)
+    )
+    private val backButtonCd = hasContentDescription(
+        androidComposeTestRule.activity.getString(R.string.sf__back_button_content_description)
+    )
+    private val deleteButtonCd = hasContentDescription(
+        androidComposeTestRule.activity.getString(R.string.sf__server_url_delete)
+    )
+
+    // region Login Server Picker Tests
 
     @OptIn(ExperimentalMaterial3Api::class)
     @Test
@@ -64,6 +83,7 @@ class PickerBottomSheetActivityTest {
             )
         }
 
+        val picker = androidComposeTestRule.onNode(serverPickerCd)
         val closeButton = androidComposeTestRule.onNode(closeButtonCd)
         val backButton = androidComposeTestRule.onNode(backButtonCd)
         val changeServerText = androidComposeTestRule.onNodeWithText(
@@ -80,6 +100,7 @@ class PickerBottomSheetActivityTest {
             androidComposeTestRule.activity.getString(R.string.sf__server_url_delete)
         )
 
+        picker.assertIsDisplayed()
         backButton.isNotDisplayed()
         changeServerText.assertIsDisplayed()
         closeButton.isDisplayed()
@@ -112,6 +133,173 @@ class PickerBottomSheetActivityTest {
 
     @OptIn(ExperimentalMaterial3Api::class)
     @Test
+    fun serverList_UpdatesOn_ServerDeletion() {
+        androidComposeTestRule.setContent {
+            PickerBottomSheetTestWrapper(
+                pickerStyle = PickerStyle.LoginServerPicker,
+                selectedListItem = serverList[1],
+            )
+        }
+
+        val prodListItem = androidComposeTestRule.onNode(prodRowCd)
+        val sandboxListItem = androidComposeTestRule.onNode(sandboxRowCd)
+        val customListItem = androidComposeTestRule.onNode(customsRowCd)
+        val customListItemRadioButton = customListItem.onChildAt(0)
+        val customListItemRemoveButton = customListItem.onChildAt(1)
+        val customListItemDeleteButton = androidComposeTestRule.onNode(deleteButtonCd)
+
+        prodListItem.onChild().assertIsNotSelected()
+        sandboxListItem.onChild().assertIsSelected()
+        customListItemRadioButton.assertIsNotSelected()
+
+        // Check only custom server has remove button
+        prodListItem.onChildAt(1).assertDoesNotExist()
+        sandboxListItem.onChildAt(1).assertDoesNotExist()
+        customListItemRemoveButton.assertExists()
+        customListItemRemoveButton.assertIsDisplayed()
+        customListItemDeleteButton.assertDoesNotExist()
+
+        // Click Remove
+        customListItemRemoveButton.performClick()
+        customListItemDeleteButton.assertExists()
+        customListItemDeleteButton.assertIsDisplayed()
+
+        // Click Delete
+        customListItemDeleteButton.performClick()
+
+        // Assert Visuals
+        prodListItem.assertIsDisplayed()
+        prodListItem.onChild().assertIsNotSelected()
+        prodListItem.assertIsDisplayed()
+        sandboxListItem.onChild().assertIsSelected()
+        customListItem.assertDoesNotExist()
+    }
+
+    @OptIn(ExperimentalMaterial3Api::class)
+    @Test
+    fun picker_DeleteServer_CallsSeverDeletion() {
+        var correctServerRemoved = false
+        val removeServer = { server: LoginServer -> if (server == customServer) correctServerRemoved = true }
+        androidComposeTestRule.setContent {
+            PickerBottomSheetTestWrapper(
+                pickerStyle = PickerStyle.LoginServerPicker,
+                selectedListItem = serverList[1],
+                removeLoginServer = removeServer,
+            )
+        }
+
+        val customListItem = androidComposeTestRule.onNode(customsRowCd)
+        val customListItemRemoveButton = customListItem.onChildAt(1)
+        val customListItemDeleteButton = androidComposeTestRule.onNode(deleteButtonCd)
+
+        // Click Remove
+        customListItemRemoveButton.performClick()
+        // Click Delete
+        customListItemDeleteButton.performClick()
+
+        // Assert Model Called
+        Assert.assertTrue("Remove server fun not called.", correctServerRemoved)
+    }
+
+    @OptIn(ExperimentalMaterial3Api::class)
+    @Test
+    fun complexServerList_UpdatesOn_ServerDeletion() {
+        val customServer2 = LoginServer("Custom2", "2mobilesdk.my.salesforce.com", true)
+        val customServer3 = LoginServer("Custom3", "3mobilesdk.my.salesforce.com", true)
+        val complexServerList = listOf(prodServer, sandboxServer, customServer, customServer2, customServer3)
+        androidComposeTestRule.setContent {
+            PickerBottomSheetTestWrapper(
+                pickerStyle = PickerStyle.LoginServerPicker,
+                list = complexServerList,
+                selectedListItem = complexServerList[4],
+            )
+        }
+
+        val prodListItem = androidComposeTestRule.onNode(prodRowCd)
+        val sandboxListItem = androidComposeTestRule.onNode(sandboxRowCd)
+        val customListItem = androidComposeTestRule.onNode(customsRowCd)
+        val customListItemRadioButton = customListItem.onChildAt(0)
+        val customListItemRemoveButton = customListItem.onChildAt(1)
+        val custom2ListItem = androidComposeTestRule.onNode(hasText(customServer2.name) and hasText(customServer2.url))
+        val custom2ListItemRadioButton = custom2ListItem.onChildAt(0)
+        val custom2ListItemRemoveButton = custom2ListItem.onChildAt(1)
+        val custom2ListItemDeleteButton = androidComposeTestRule.onNode(deleteButtonCd)
+        val custom3ListItem = androidComposeTestRule.onNode(hasText(customServer3.name) and hasText(customServer3.url))
+        val custom3ListItemRadioButton = custom3ListItem.onChildAt(0)
+        val custom3ListItemRemoveButton = custom3ListItem.onChildAt(1)
+
+        prodListItem.onChild().assertIsNotSelected()
+        sandboxListItem.onChild().assertIsNotSelected()
+        customListItemRadioButton.assertIsNotSelected()
+        custom2ListItemRadioButton.assertIsNotSelected()
+        custom3ListItemRadioButton.assertIsSelected()
+
+        // Check only custom servers have remove button
+        prodListItem.onChildAt(1).assertDoesNotExist()
+        sandboxListItem.onChildAt(1).assertDoesNotExist()
+        customListItemRemoveButton.assertIsDisplayed()
+        custom2ListItemRemoveButton.assertIsDisplayed()
+        custom2ListItemDeleteButton.assertDoesNotExist()
+        custom3ListItemRemoveButton.assertIsDisplayed()
+
+        // Click Remove
+        custom2ListItemRemoveButton.performClick()
+        custom2ListItemDeleteButton.assertExists()
+        custom2ListItemDeleteButton.assertIsDisplayed()
+
+        // Click Delete
+        custom2ListItemDeleteButton.performClick()
+
+        // Assert Visuals
+        prodListItem.assertIsDisplayed()
+        prodListItem.assertIsDisplayed()
+        customListItem.assertIsDisplayed()
+        custom2ListItem.assertDoesNotExist()
+        custom3ListItem.assertIsDisplayed()
+    }
+
+    @OptIn(ExperimentalMaterial3Api::class)
+    @Test
+    fun serverList_UpdatesOn_SelectedServerDeletion() {
+        androidComposeTestRule.setContent {
+            PickerBottomSheetTestWrapper(
+                pickerStyle = PickerStyle.LoginServerPicker,
+                selectedListItem = serverList[2],
+            )
+        }
+
+        val prodListItem = androidComposeTestRule.onNode(prodRowCd)
+        val sandboxListItem = androidComposeTestRule.onNode(sandboxRowCd)
+        val customListItem = androidComposeTestRule.onNode(customsRowCd)
+        val customListItemRadioButton = customListItem.onChildAt(0)
+        val customListItemRemoveButton = customListItem.onChildAt(1)
+        val customListItemDeleteButton = androidComposeTestRule.onNode(deleteButtonCd)
+
+        prodListItem.onChild().assertIsNotSelected()
+        sandboxListItem.onChild().assertIsNotSelected()
+        customListItemRadioButton.assertIsSelected()
+
+        // Click Remove
+        customListItemRemoveButton.performClick()
+        customListItemDeleteButton.assertExists()
+        customListItemDeleteButton.assertIsDisplayed()
+
+        // Click Delete
+        customListItemDeleteButton.performClick()
+
+        // Assert Visuals - First server should now be selected.
+        prodListItem.assertIsDisplayed()
+        prodListItem.onChild().assertIsSelected()
+        prodListItem.assertIsDisplayed()
+        sandboxListItem.onChild().assertIsNotSelected()
+        customListItem.assertDoesNotExist()
+    }
+
+    // endregion
+    // region Account Picker Tests
+
+    @OptIn(ExperimentalMaterial3Api::class)
+    @Test
     fun accountPickerHeader_Displays_Correctly() {
         androidComposeTestRule.setContent {
             PickerBottomSheetTestWrapper(
@@ -120,6 +308,7 @@ class PickerBottomSheetActivityTest {
             )
         }
 
+        val picker = androidComposeTestRule.onNode(accountPickerCd)
         val closeButton = androidComposeTestRule.onNode(closeButtonCd)
         val backButton = androidComposeTestRule.onNode(backButtonCd)
         val accountSelectorText = androidComposeTestRule.onNodeWithText(
@@ -129,9 +318,34 @@ class PickerBottomSheetActivityTest {
             androidComposeTestRule.activity.getString(R.string.sf__add_new_account)
         )
 
+        picker.assertIsDisplayed()
         backButton.assertDoesNotExist()
         accountSelectorText.assertIsDisplayed()
         closeButton.isDisplayed()
         addAccountButton.isDisplayed()
     }
+
+    @OptIn(ExperimentalMaterial3Api::class)
+    @Test
+    fun userPicker_NewAccountButtonClick_CallsAddNewAccount() {
+        var addAccountCalled = false
+        androidComposeTestRule.setContent {
+            PickerBottomSheetTestWrapper(
+                pickerStyle = PickerStyle.UserAccountPicker,
+                addNewAccount = { addAccountCalled = true },
+            )
+        }
+
+        val addNewAccountButton = androidComposeTestRule.onNodeWithText(
+            androidComposeTestRule.activity.getString(R.string.sf__add_new_account)
+        )
+
+        // Close Picker
+        addNewAccountButton.performClick()
+
+        // Assert Cancel Called
+        Assert.assertTrue("addNewAccount not called.", addAccountCalled)
+    }
+
+    // endregion
 }
