+++ libs/test/SalesforceSDKTest/src/com/salesforce/androidsdk/ui/PickerBottomSheetTest.kt
@@ -39,7 +39,6 @@ import androidx.compose.ui.test.assertIsNotEnabled
 import androidx.compose.ui.test.assertIsNotSelected
 import androidx.compose.ui.test.assertIsSelectable
 import androidx.compose.ui.test.assertIsSelected
-import androidx.compose.ui.test.hasContentDescription
 import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onChild
@@ -49,15 +48,9 @@ import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextClearance
 import androidx.compose.ui.test.performTextInput
 import com.salesforce.androidsdk.config.LoginServerManager.LoginServer
-import com.salesforce.androidsdk.ui.components.ADD_NEW_BUTTON_CD
 import com.salesforce.androidsdk.ui.components.AddConnection
-import com.salesforce.androidsdk.ui.components.CLOSE_BUTTON_CD
-import com.salesforce.androidsdk.ui.components.DELETE_BUTTON_CD
-import com.salesforce.androidsdk.ui.components.LOGIN_SERVER_CD
-import com.salesforce.androidsdk.ui.components.PICKER_CD
 import com.salesforce.androidsdk.ui.components.PickerBottomSheet
 import com.salesforce.androidsdk.ui.components.PickerStyle
-import com.salesforce.androidsdk.ui.components.USER_ACCOUNT_CD
 import com.salesforce.androidsdk.ui.components.UserAccountMock
 import org.junit.Assert
 import org.junit.Rule
@@ -71,18 +64,25 @@ private const val VALID_URL = "https://login.salesforce.com"
 private const val INVALID_URL = "invalid"
 private const val WITH_VALIDATION = "_with_validation"
 
-private val prodServer = LoginServer(TEST_NAME, VALID_URL, false)
-private val sandboxServer = LoginServer("Sandbox", "https://test.salesforce.com", false)
-private val customServer = LoginServer("Custom", "https://mobilesdk.my.salesforce.com", true)
-private val serverList = listOf(prodServer, sandboxServer, customServer)
+@VisibleForTesting
+internal val prodServer = LoginServer(TEST_NAME, VALID_URL, false)
+@VisibleForTesting
+internal val sandboxServer = LoginServer("Sandbox", "https://test.salesforce.com", false)
+@VisibleForTesting
+internal val customServer = LoginServer("Custom", "https://mobilesdk.my.salesforce.com", true)
+@VisibleForTesting
+internal val serverList = listOf(prodServer, sandboxServer, customServer)
+@VisibleForTesting
+internal val prodRowCd = (hasText(prodServer.name) and hasText(prodServer.url))
+@VisibleForTesting
+internal val sandboxRowCd = (hasText(sandboxServer.name) and hasText(sandboxServer.url))
 
 private val user1 = UserAccountMock("user1", VALID_URL)
 private val user2 = UserAccountMock("user2", sandboxServer.url)
 private val userList = listOf(user1, user2)
 
 @VisibleForTesting
-internal val customsRowCd = (hasContentDescription(LOGIN_SERVER_CD)
-        and hasText(customServer.name) and hasText(customServer.url))
+internal val customsRowCd = (hasText(customServer.name) and hasText(customServer.url))
 
 class PickerBottomSheetTest {
 
@@ -91,18 +91,9 @@ class PickerBottomSheetTest {
 
     /* This call will print the semantic tree: composeTestRule.onAllNodes(isRoot()).printToLog("", 10) */
 
-    private val pickerCd = hasContentDescription(PICKER_CD)
-    private val closeButtonCd = hasContentDescription(CLOSE_BUTTON_CD)
-    private val prodRowCd = (hasContentDescription(LOGIN_SERVER_CD)
-            and hasText(prodServer.name) and hasText(prodServer.url))
-    private val sandboxRowCd = (hasContentDescription(LOGIN_SERVER_CD)
-            and hasText(sandboxServer.name) and hasText(sandboxServer.url))
-    private val customsRowCd = (hasContentDescription(LOGIN_SERVER_CD)
-            and hasText(customServer.name) and hasText(customServer.url))
-    private val user1RowCd = (hasContentDescription(USER_ACCOUNT_CD)
-            and hasText(user1.displayName) and hasText(user1.loginServer))
-    private val user2RowCd = (hasContentDescription(USER_ACCOUNT_CD)
-            and hasText(user2.displayName) and hasText(user2.loginServer))
+    private val customsRowCd = (hasText(customServer.name) and hasText(customServer.url))
+    private val user1RowCd = (hasText(user1.displayName) and hasText(user1.loginServer))
+    private val user2RowCd = (hasText(user2.displayName) and hasText(user2.loginServer))
 
     // Google's recommended naming scheme for test is "thingUnderTest_TriggerOfTest_ResultOfTest"
 
@@ -182,14 +173,10 @@ class PickerBottomSheetTest {
             )
         }
 
-        val picker = composeTestRule.onNode(pickerCd)
-        val closeButton = composeTestRule.onNode(closeButtonCd)
         val prodListItem = composeTestRule.onNode(prodRowCd)
         val sandboxListItem = composeTestRule.onNode(sandboxRowCd)
         val customListItem = composeTestRule.onNode(customsRowCd)
 
-        picker.assertIsDisplayed()
-        closeButton.assertIsDisplayed()
         prodListItem.assertIsDisplayed()
         prodListItem.onChild().assertIsSelectable()
         prodListItem.onChild().assertIsSelected()
@@ -287,172 +274,6 @@ class PickerBottomSheetTest {
         Assert.assertEquals(sandboxServer, selectedServer)
     }
 
-    @OptIn(ExperimentalMaterial3Api::class)
-    @Test
-    fun serverList_UpdatesOn_ServerDeletion() {
-        composeTestRule.setContent {
-            PickerBottomSheetTestWrapper(
-                pickerStyle = PickerStyle.LoginServerPicker,
-                selectedListItem = serverList[1],
-            )
-        }
-
-        val prodListItem = composeTestRule.onNode(prodRowCd)
-        val sandboxListItem = composeTestRule.onNode(sandboxRowCd)
-        val customListItem = composeTestRule.onNode(customsRowCd)
-        val customListItemRadioButton = customListItem.onChildAt(0)
-        val customListItemRemoveButton = customListItem.onChildAt(1)
-        val customListItemDeleteButton = composeTestRule.onNode(hasContentDescription(DELETE_BUTTON_CD))
-
-        prodListItem.onChild().assertIsNotSelected()
-        sandboxListItem.onChild().assertIsSelected()
-        customListItemRadioButton.assertIsNotSelected()
-
-        // Check only custom server has remove button
-        prodListItem.onChildAt(1).assertDoesNotExist()
-        sandboxListItem.onChildAt(1).assertDoesNotExist()
-        customListItemRemoveButton.assertExists()
-        customListItemRemoveButton.assertIsDisplayed()
-        customListItemDeleteButton.assertDoesNotExist()
-
-        // Click Remove
-        customListItemRemoveButton.performClick()
-        customListItemDeleteButton.assertExists()
-        customListItemDeleteButton.assertIsDisplayed()
-
-        // Click Delete
-        customListItemDeleteButton.performClick()
-
-        // Assert Visuals
-        prodListItem.assertIsDisplayed()
-        prodListItem.onChild().assertIsNotSelected()
-        prodListItem.assertIsDisplayed()
-        sandboxListItem.onChild().assertIsSelected()
-        customListItem.assertDoesNotExist()
-    }
-
-    @OptIn(ExperimentalMaterial3Api::class)
-    @Test
-    fun picker_DeleteServer_CallsSeverDeletion() {
-        var correctServerRemoved = false
-        val removeServer = { server: LoginServer -> if (server == customServer) correctServerRemoved = true }
-        composeTestRule.setContent {
-            PickerBottomSheetTestWrapper(
-                pickerStyle = PickerStyle.LoginServerPicker,
-                selectedListItem = serverList[1],
-                removeLoginServer = removeServer,
-            )
-        }
-
-        val customListItem = composeTestRule.onNode(customsRowCd)
-        val customListItemRemoveButton = customListItem.onChildAt(1)
-        val customListItemDeleteButton = composeTestRule.onNode(hasContentDescription(DELETE_BUTTON_CD))
-
-        // Click Remove
-        customListItemRemoveButton.performClick()
-        // Click Delete
-        customListItemDeleteButton.performClick()
-
-        // Assert Model Called
-        Assert.assertTrue("Remove server fun not called.", correctServerRemoved)
-    }
-
-    @OptIn(ExperimentalMaterial3Api::class)
-    @Test
-    fun complexServerList_UpdatesOn_ServerDeletion() {
-        val customServer2 = LoginServer("Custom2", "2mobilesdk.my.salesforce.com", true)
-        val customServer3 = LoginServer("Custom3", "3mobilesdk.my.salesforce.com", true)
-        val complexServerList = listOf(prodServer, sandboxServer, customServer, customServer2, customServer3)
-        composeTestRule.setContent {
-            PickerBottomSheetTestWrapper(
-                pickerStyle = PickerStyle.LoginServerPicker,
-                list = complexServerList,
-                selectedListItem = complexServerList[4],
-            )
-        }
-
-        val prodListItem = composeTestRule.onNode(prodRowCd)
-        val sandboxListItem = composeTestRule.onNode(sandboxRowCd)
-        val customListItem = composeTestRule.onNode(customsRowCd)
-        val customListItemRadioButton = customListItem.onChildAt(0)
-        val customListItemRemoveButton = customListItem.onChildAt(1)
-        val custom2ListItem = composeTestRule.onNode((hasContentDescription(LOGIN_SERVER_CD)
-                and hasText(customServer2.name) and hasText(customServer2.url)))
-        val custom2ListItemRadioButton = custom2ListItem.onChildAt(0)
-        val custom2ListItemRemoveButton = custom2ListItem.onChildAt(1)
-        val custom2ListItemDeleteButton = composeTestRule.onNode(hasContentDescription(DELETE_BUTTON_CD))
-        val custom3ListItem = composeTestRule.onNode((hasContentDescription(LOGIN_SERVER_CD)
-                and hasText(customServer3.name) and hasText(customServer3.url)))
-        val custom3ListItemRadioButton = custom3ListItem.onChildAt(0)
-        val custom3ListItemRemoveButton = custom3ListItem.onChildAt(1)
-
-        prodListItem.onChild().assertIsNotSelected()
-        sandboxListItem.onChild().assertIsNotSelected()
-        customListItemRadioButton.assertIsNotSelected()
-        custom2ListItemRadioButton.assertIsNotSelected()
-        custom3ListItemRadioButton.assertIsSelected()
-
-        // Check only custom servers have remove button
-        prodListItem.onChildAt(1).assertDoesNotExist()
-        sandboxListItem.onChildAt(1).assertDoesNotExist()
-        customListItemRemoveButton.assertIsDisplayed()
-        custom2ListItemRemoveButton.assertIsDisplayed()
-        custom2ListItemDeleteButton.assertDoesNotExist()
-        custom3ListItemRemoveButton.assertIsDisplayed()
-
-        // Click Remove
-        custom2ListItemRemoveButton.performClick()
-        custom2ListItemDeleteButton.assertExists()
-        custom2ListItemDeleteButton.assertIsDisplayed()
-
-        // Click Delete
-        custom2ListItemDeleteButton.performClick()
-
-        // Assert Visuals
-        prodListItem.assertIsDisplayed()
-        prodListItem.assertIsDisplayed()
-        customListItem.assertIsDisplayed()
-        custom2ListItem.assertDoesNotExist()
-        custom3ListItem.assertIsDisplayed()
-    }
-
-    @OptIn(ExperimentalMaterial3Api::class)
-    @Test
-    fun serverList_UpdatesOn_SelectedServerDeletion() {
-        composeTestRule.setContent {
-            PickerBottomSheetTestWrapper(
-                pickerStyle = PickerStyle.LoginServerPicker,
-                selectedListItem = serverList[2],
-            )
-        }
-
-        val prodListItem = composeTestRule.onNode(prodRowCd)
-        val sandboxListItem = composeTestRule.onNode(sandboxRowCd)
-        val customListItem = composeTestRule.onNode(customsRowCd)
-        val customListItemRadioButton = customListItem.onChildAt(0)
-        val customListItemRemoveButton = customListItem.onChildAt(1)
-        val customListItemDeleteButton = composeTestRule.onNode(hasContentDescription(DELETE_BUTTON_CD))
-
-        prodListItem.onChild().assertIsNotSelected()
-        sandboxListItem.onChild().assertIsNotSelected()
-        customListItemRadioButton.assertIsSelected()
-
-        // Click Remove
-        customListItemRemoveButton.performClick()
-        customListItemDeleteButton.assertExists()
-        customListItemDeleteButton.assertIsDisplayed()
-
-        // Click Delete
-        customListItemDeleteButton.performClick()
-
-        // Assert Visuals - First server should now be selected.
-        prodListItem.assertIsDisplayed()
-        prodListItem.onChild().assertIsSelected()
-        prodListItem.assertIsDisplayed()
-        sandboxListItem.onChild().assertIsNotSelected()
-        customListItem.assertDoesNotExist()
-    }
-
     // endregion
     // region Account Picker Tests
 
@@ -466,13 +287,9 @@ class PickerBottomSheetTest {
             )
         }
 
-        val picker = composeTestRule.onNode(pickerCd)
-        val closeButton = composeTestRule.onNode(closeButtonCd)
         val user1ListItem = composeTestRule.onNode(user1RowCd)
         val user2ListItem = composeTestRule.onNode(user2RowCd)
 
-        picker.assertIsDisplayed()
-        closeButton.assertIsDisplayed()
         user1ListItem.assertIsDisplayed()
         user1ListItem.onChild().assertIsSelectable()
         user1ListItem.onChild().assertIsSelected()
@@ -536,26 +353,6 @@ class PickerBottomSheetTest {
 
         // Assert data source change
         Assert.assertEquals(user2, currentUser)
-    }
-
-    @OptIn(ExperimentalMaterial3Api::class)
-    @Test
-    fun userPicker_NewAccountButtonClick_CallsAddNewAccount() {
-        var addAccountCalled = false
-        composeTestRule.setContent {
-            PickerBottomSheetTestWrapper(
-                pickerStyle = PickerStyle.UserAccountPicker,
-                addNewAccount = { addAccountCalled = true },
-            )
-        }
-
-        val addNewAccountButton = composeTestRule.onNode(hasContentDescription(ADD_NEW_BUTTON_CD))
-
-        // Close Picker
-        addNewAccountButton.performClick()
-
-        // Assert Cancel Called
-        Assert.assertTrue("addNewAccount not called.", addAccountCalled)
     }
 
     // endregion
